US122289 Renaming components to reflect current roles
