+++ test/components/message-container.test.js
@@ -1,36 +1,9 @@
 import '../../components/message-container';
 
-import { expect, fixture, html } from '@open-wc/testing';
+import { expect, fixture, html, oneEvent } from '@open-wc/testing';
 import { runConstructor } from '@brightspace-ui/core/tools/constructor-test-helper.js';
 
 describe('d2l-insights-message-container', () => {
-
-	const dataWithTruncatedRecords = {
-		_data: {
-			serverData : {
-				isRecordsTruncated: true
-			}
-		},
-		users: [
-			[100, 'John', 'Lennon', 'jlennon',  1600295350000],
-			[200, 'Paul', 'McCartney', 'pmccartney', null]
-		]
-	};
-
-	const dataWithoutTruncatedRecords = {
-		_data: {
-			serverData : {
-				isRecordsTruncated: false
-			},
-		},
-		users: [],
-	};
-
-	const dataWithQueryError = {
-		_data: {
-			isQueryError: true
-		}
-	};
 
 	describe('constructor', () => {
 		it('should construct', () => {
@@ -38,41 +11,72 @@ describe('d2l-insights-message-container', () => {
 		});
 	});
 
-	describe('accessibility', () => {
+	describe('default message', () => {
+		let el;
+		before(async() => {
+			el = await fixture(html`
+				<d2l-insights-message-container type="default" text="some text"></d2l-insights-message-container>
+			`);
+		});
+
 		it('should pass all axe tests', async() => {
-			const el = await fixture(html`<d2l-insights-message-container .data="${dataWithTruncatedRecords}" .isNoDataReturned="${Boolean(0)}"></d2l-insights-message-container>`);
 			await expect(el).to.be.accessible();
+		});
+
+		it('should render correctly', () => {
+			const innerSpan = el.shadowRoot.querySelector('span');
+			expect(innerSpan.innerText).to.equal('some text');
 		});
 	});
 
-	describe('render', () => {
-		it('should render as expected with truncated records', async() => {
-			const el = await fixture(html`<d2l-insights-message-container .data="${dataWithTruncatedRecords}" .isNoDataReturned="${Boolean(0)}"></d2l-insights-message-container>`);
-			const innerMessageContainer = el.shadowRoot.querySelector('d2l-insights-error-message');
-			expect(innerMessageContainer.type).to.equal('default');
-			expect(innerMessageContainer.text).to.equal('There are too many results in your filters. Please refine your selection.');
+	describe('message with button', () => {
+		let el;
+		before(async() => {
+			el = await fixture(html`
+				<d2l-insights-message-container type="button" text="some text" button-text="button text"></d2l-insights-message-container>
+			`);
 		});
 
-		it('should not render without truncated records', async() => {
-			const el = await fixture(html`<d2l-insights-message-container .data="${dataWithoutTruncatedRecords}" .isNoDataReturned="${Boolean(1)}"></d2l-insights-message-container>`);
-			const innerMessageContainer = el.shadowRoot.querySelector('d2l-insights-error-message');
-			expect(innerMessageContainer.type).to.equal('button');
-			expect(innerMessageContainer.text).to.equal('There are no results available that match your filters.');
-			expect(innerMessageContainer.buttonText).to.equal('Undo Last Action');
+		it('should pass all axe tests', async() => {
+			await expect(el).to.be.accessible();
 		});
 
-		it('should render as expected with contact support message', async() => {
-			const el = await fixture(html`<d2l-insights-message-container .data="${dataWithQueryError}" .isNoDataReturned="${Boolean(0)}"></d2l-insights-message-container>`);
-			const innerMessageContainer = el.shadowRoot.querySelector('d2l-insights-error-message');
-			expect(innerMessageContainer.type).to.equal('link');
-			expect(innerMessageContainer.text).to.equal('Unable to load your results. If this problem persists, please ');
-			expect(innerMessageContainer.linkText).to.equal('contact D2L Support.');
+		it('should render correctly', () => {
+			const innerSpan = el.shadowRoot.querySelector('span');
+			expect(innerSpan.innerText).to.equal('some text');
+
+			const button = el.shadowRoot.querySelector('d2l-button');
+			expect(button.innerText).to.equal('button text');
 		});
 
-		it('should render nothing if there are no issues', async() => {
-			const el = await fixture(html`<d2l-insights-message-container .data="${dataWithoutTruncatedRecords}" .isNoDataReturned="${Boolean(0)}"></d2l-insights-message-container>`);
-			const innerMessageContainer = el.shadowRoot.querySelector('d2l-insights-error-message');
-			expect(innerMessageContainer).to.be.null;
+		it('should fire an event when clicked', async() => {
+			const listener = oneEvent(el, 'd2l-insights-message-container-button-click');
+			const button = el.shadowRoot.querySelector('d2l-button');
+			button.click();
+			await listener;
+		});
+	});
+
+	describe('message with link', () => {
+		let el;
+		before(async() => {
+			el = await fixture(html`
+				<d2l-insights-message-container type="link" text="some text" href="https://www.example.com" link-text="link text"></d2l-insights-message-container>
+			`);
+		});
+
+		it('should pass all axe tests', async() => {
+			await expect(el).to.be.accessible();
+		});
+
+		it('should render correctly', () => {
+			const innerSpan = el.shadowRoot.querySelector('span');
+			// it adds a bunch of \n and \t characters which don't appear on-screen
+			expect(innerSpan.innerText.trim().split('\n')[0]).to.equal('some text');
+
+			const link = el.shadowRoot.querySelector('a');
+			expect(link.innerText).to.equal('link text');
+			expect(link.href).to.equal('https://www.example.com/');
 		});
 	});
 });
